added typeFile reader function to dynamically retrieve method types where necessary
AppUserType changes over time of development, to not always have to update this appwrite-wrapper library, i made some kind of model/migrate so that a user type will be created based on a json schema about this type. This will add a collection with the necessary attributes as well as creating a type file on the users projects folder ("lib/appwrite/schemas/<collectionName>.ts"). And this file now gets read in using the newly created typeFile reader to cast exports from the respectie functions that use it